Make ValidateApplicationService work with UserIDs and move to internal
