--- appservice/validate/validate_test.go
@@ -1,86 +1,0 @@
-// Copyright 2023 The Matrix.org Foundation C.I.C.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package validate
-
-import (
-	"regexp"
-	"testing"
-
-	"github.com/matrix-org/dendrite/setup/config"
-)
-
-// This method tests validation of the provided Application Service token and
-// username that they're registering
-func TestValidationOfApplicationServices(t *testing.T) {
-	// Set up application service namespaces
-	regex := "@_appservice_.*"
-	regexp, err := regexp.Compile(regex)
-	if err != nil {
-		t.Errorf("Error compiling regex: %s", regex)
-	}
-
-	fakeNamespace := config.ApplicationServiceNamespace{
-		Exclusive:    true,
-		Regex:        regex,
-		RegexpObject: regexp,
-	}
-
-	// Create a fake application service
-	fakeID := "FakeAS"
-	fakeSenderLocalpart := "_appservice_bot"
-	fakeApplicationService := config.ApplicationService{
-		ID:              fakeID,
-		URL:             "null",
-		ASToken:         "1234",
-		HSToken:         "4321",
-		SenderLocalpart: fakeSenderLocalpart,
-		NamespaceMap: map[string][]config.ApplicationServiceNamespace{
-			"users": {fakeNamespace},
-		},
-	}
-
-	// Set up a config
-	fakeConfig := &config.Dendrite{}
-	fakeConfig.Defaults(config.DefaultOpts{
-		Generate:   true,
-		Monolithic: true,
-	})
-	fakeConfig.Global.ServerName = "localhost"
-	fakeConfig.ClientAPI.Derived.ApplicationServices = []config.ApplicationService{fakeApplicationService}
-
-	// Access token is correct, user_id omitted so we are acting as SenderLocalpart
-	asID, resp := ValidateApplicationService(&fakeConfig.ClientAPI, fakeSenderLocalpart, "1234")
-	if resp != nil || asID != fakeID {
-		t.Errorf("appservice should have validated and returned correct ID: %s", resp.JSON)
-	}
-
-	// Access token is incorrect, user_id omitted so we are acting as SenderLocalpart
-	asID, resp = ValidateApplicationService(&fakeConfig.ClientAPI, fakeSenderLocalpart, "xxxx")
-	if resp == nil || asID == fakeID {
-		t.Errorf("access_token should have been marked as invalid")
-	}
-
-	// Access token is correct, acting as valid user_id
-	asID, resp = ValidateApplicationService(&fakeConfig.ClientAPI, "_appservice_bob", "1234")
-	if resp != nil || asID != fakeID {
-		t.Errorf("access_token and user_id should've been valid: %s", resp.JSON)
-	}
-
-	// Access token is correct, acting as invalid user_id
-	asID, resp = ValidateApplicationService(&fakeConfig.ClientAPI, "_something_else", "1234")
-	if resp == nil || asID == fakeID {
-		t.Errorf("user_id should not have been valid: @_something_else:localhost")
-	}
-}
